Add ViewPort message when starting to record
In order for replays to work the view port must be set before any mouse
and keyboard events are sent.   Capture the view port and send it to the
server when the recording is started.   This will send the view port
parameters to the UI as the first command when the recording is
replayed.